+++ Apps/WebClient/src/ClientApp/src/models/requestResult.ts
@@ -1,5 +1,5 @@
+import { ActionType } from "@/constants/actionType";
 import { ResultType } from "@/constants/resulttype";
-import { ResultErrorDetails } from "@/models/errors";
 
 export default interface RequestResult<T> {
     // The request resource payload
@@ -14,4 +14,18 @@ export default interface RequestResult<T> {
     resultStatus: ResultType;
     // The result error associated to the request (could be empty)
     resultError?: ResultErrorDetails;
+}
+
+// HG API binding interface for ErrorResult model
+export interface ResultErrorDetails {
+    // Message that will always be populated when ResultType is Error.
+    resultMessage?: string;
+    // Code that will always be populated when ResultType is Error.
+    errorCode: string;
+    // The trace ID associated with the request.
+    traceId: string;
+    // The action code that will be set when ResultType is ActionRequired.
+    actionCode?: ActionType;
+    // The HTTP status code returned by the request.
+    statusCode?: number;
 }
