+++ Apps/WebClient/src/ClientApp/src/stores/clinicalDocument.ts
@@ -128,9 +128,7 @@ export const useClinicalDocumentStore = defineStore("clinicalDocument", () => {
                     );
                 } else {
                     if (result.resultError) {
-                        throw ResultError.fromResultErrorDetails(
-                            result.resultError
-                        );
+                        throw ResultError.fromModel(result.resultError);
                     }
                     logger.warn(
                         `Clinical documents retrieval failed! ${JSON.stringify(
@@ -162,9 +160,7 @@ export const useClinicalDocumentStore = defineStore("clinicalDocument", () => {
                     setFile(fileId, payload);
                 } else {
                     if (result.resultError) {
-                        throw ResultError.fromResultErrorDetails(
-                            result.resultError
-                        );
+                        throw ResultError.fromModel(result.resultError);
                     }
                     logger.warn(
                         `Clinical document file retrieval failed! ${JSON.stringify(
