+++ Apps/WebClient/src/ClientApp/src/stores/vaccinationStatusAuthenticated.ts
@@ -275,9 +275,7 @@ export const useVaccinationStatusAuthenticatedStore = defineStore(
                             }, payload.retryin);
                         } else {
                             if (result.resultError) {
-                                throw ResultError.fromResultErrorDetails(
-                                    result.resultError
-                                );
+                                throw ResultError.fromModel(result.resultError);
                             }
                             logger.warn(
                                 `Authenticated vaccination status retrieval failed! ${JSON.stringify(
@@ -324,9 +322,7 @@ export const useVaccinationStatusAuthenticatedStore = defineStore(
                         }, payload.retryin);
                     } else {
                         if (result.resultError) {
-                            throw ResultError.fromResultErrorDetails(
-                                result.resultError
-                            );
+                            throw ResultError.fromModel(result.resultError);
                         }
                         logger.warn(
                             `Authenticated vaccination record retrieval failed! ${JSON.stringify(
